Desmembrando o programa MAIN para passar as estruturas para arquivos separados
