+++ src/allegro.h
@@ -1,0 +1,9 @@
+#ifndef ALLEGRO_H
+#define ALLEGRO_H
+
+#include <allegro5/allegro.h>
+#include <allegro5/allegro_image.h>
+#include <allegro5/allegro_font.h>   
+#include <allegro5/allegro_ttf.h>
+
+#endif
