+++ src/inicializa.h
@@ -1,0 +1,7 @@
+#ifndef INICIALIZA_H
+#define INICIALIZA_H
+
+int inicializa();
+int desinicializa();
+
+#endif
