+++ src/menu.cpp
@@ -1,0 +1,40 @@
+#include "allegro.h"
+#include "main.h"
+#include "menu.h"
+
+int mainmenu(){
+    /*
+        ESTA ESTRUTURA PERMITE O JOGO INICIALIZAR UMA FILA DE EVENTOS QUE PODE SER QUEBRADA SEM DESLIGAR O JOGO
+        EXEMPLO: O jogador inicia o tabuleiro e quando mandamos um BREAK na função jogo(), o programa retorna ao menu inicial, por conta da função Main, que está em um loop
+    */
+    ALLEGRO_EVENT event;                        //INICIA UMA ÚNICA FILA DE EVENTOS
+    al_draw_bitmap(menu,0,0,0);
+    al_flip_display();
+    al_wait_for_event(event_queue, &event);     //ESPERA POR UM EVENTO (No caso da condição abaixo, iniciará o jogo). SE FOR QUEBRADO(receber um break), A CONDIÇÃO RETORNA O PROGRAMA AO MENU INICIAR
+    if(event.type == ALLEGRO_EVENT_KEY_UP && event.keyboard.keycode==ALLEGRO_KEY_UP){
+        if(keyboardState<3){
+            keyboardState += 1;
+        }
+    }
+    if(event.type == ALLEGRO_EVENT_KEY_UP && event.keyboard.keycode==ALLEGRO_KEY_DOWN){
+        if(keyboardState>0){
+            keyboardState -= 1;
+        }
+    }
+    if(event.type == ALLEGRO_EVENT_MOUSE_BUTTON_DOWN && event.mouse.button == 1){
+        if(keyboardState == 1){
+            jogo(event);
+        }
+    }
+    //if keyboardState == 2 { segundaFase() }
+    //if keyboardState == 3 { terceiraFase() }
+    //if keyboardState == 4 { quartaFase() }
+
+    else if(event.type == ALLEGRO_EVENT_KEY_UP){
+        if(event.keyboard.keycode==ALLEGRO_KEY_ESCAPE){
+                sair=true;
+                return sair;
+        }
+    }
+    
+}
