+++ src/menu.h
@@ -1,6 +1,9 @@
 #ifndef MENU_H
 #define MENU_H
 
+
+int mainmenu();
+
 class Menu {
     private:
      ////////////////////////////////
